# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/budgets": {
      "get": {
        "operationId": "BudgetController_listBudgets",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List organization budgets",
        "tags": ["Budgets and usage"]
      },
      "post": {
        "operationId": "BudgetController_createBudget",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create an organization budget",
        "tags": ["Budgets and usage"]
      }
    },
    "/api/v1/organizations/{organizationId}/usage": {
      "get": {
        "operationId": "BudgetController_usage",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List usage records for the organization",
        "tags": ["Budgets and usage"]
      }
    },
    "/api/v1/organizations/{organizationId}/pricing/models": {
      "post": {
        "operationId": "BudgetController_createPricing",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a versioned model pricing record",
        "tags": ["Budgets and usage"]
      }
    },
    "/api/v1/organizations/{organizationId}/capabilities": {
      "get": {
        "operationId": "CapabilityController_list",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List capabilities available on the API channel",
        "tags": ["Capabilities"]
      }
    },
    "/api/v1/organizations/{organizationId}/capabilities/{slug}": {
      "get": {
        "operationId": "CapabilityController_get",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resolve an API capability",
        "tags": ["Capabilities"]
      }
    },
    "/api/v1/organizations/{organizationId}/capabilities/{slug}/run": {
      "post": {
        "operationId": "CapabilityController_run",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Execute a capability through the governed engine",
        "tags": ["Capabilities"]
      }
    },
...
    "/api/v1/organizations/{organizationId}/access-requests": {
      "get": {
        "operationId": "AccessController_list",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List access requests",
        "tags": ["Access governance"]
      },
      "post": {
        "operationId": "AccessController_submit",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit an access request",
        "tags": ["Access governance"]
      }
    },
    "/api/v1/organizations/{organizationId}/access-requests/{requestId}/approve": {
      "post": {
        "operationId": "AccessController_approve",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "requestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve an access request",
        "tags": ["Access governance"]
      }
    },
    "/api/v1/organizations/{organizationId}/access-requests/{requestId}/grants/{grantId}/revoke": {
      "post": {
        "operationId": "AccessController_revoke",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "grantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke an access grant",
        "tags": ["Access governance"]
      }
    },
...
  },
  "info": {
    "title": "HandStack API",
    "description": "Versioned public and administrative HandStack API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}